Restructured FAIRMODE stats implementation and tests
